[HS] Start Homework 2, Exercise 1
The first intuition that came to my head was to model parsing after
definite clause grammars in Prolog. Function composition is kind of ugly
because I don't know an elegant way to compose `a -> (a, ?)`, so I
resort to nested pattern matching.

My intuition thinks that I can use the state monad to define the clauses
and use Kleisli composition, but I don't know enough to try that idea.